refactor(fantasy): move query hooks + centralize invalidation + barrel export
Move React Query hooks (events, fantasyLeagues, fantasyPicker) from
src/lib/queries/ into src/features/fantasy/queries/. Create centralized
invalidation functions (invalidateAfterJoin, invalidateAfterLeave,
invalidateAfterLineupSave, invalidateAfterScoring). Old files become
re-export bridges for backward compatibility. Feature barrel export
at src/features/fantasy/index.ts.
