+++ src/features/fantasy/index.ts
@@ -1,0 +1,23 @@
+// Types
+export * from './types';
+export * from './constants';
+export * from './helpers';
+
+// Mappers
+export { deriveEventStatus, dbEventToFantasyEvent } from './mappers/eventMapper';
+export { dbHoldingToUserDpcHolding } from './mappers/holdingMapper';
+
+// Services — Public API (only commonly used functions)
+export { getEvents, getEventsByClubId, getUserJoinedEventIds, isClubEvent } from './services/events.queries';
+export { lockEventEntry, unlockEventEntry } from './services/events.mutations';
+export { getLineup, getOwnedPlayerIds, getEventParticipants, getEventParticipantCount } from './services/lineups.queries';
+export type { LineupSlotPlayer, LineupWithPlayers } from './services/lineups.queries';
+export { submitLineup } from './services/lineups.mutations';
+export { getEventLeaderboard, getProgressiveScores } from './services/scoring.queries';
+export type { ScoreResult, LeaderboardEntry } from './services/scoring.queries';
+
+// Query Hooks
+export { useEvents, useJoinedEventIds, usePlayerEventUsage, useHoldingLocks, useWildcardBalance, useLeagueActiveGameweek, useIsClubAdmin, useEventEntry, useEnteredEventIds, useScoutEventsEnabled } from './queries/events';
+
+// Invalidation
+export { invalidateAfterJoin, invalidateAfterLeave, invalidateAfterLineupSave, invalidateAfterScoring } from './queries/invalidation';
