+++ src/features/fantasy/queries/fantasyLeagues.ts
@@ -1,0 +1,20 @@
+import { useQuery } from '@tanstack/react-query';
+import { getMyLeagues, getLeagueLeaderboard } from '@/features/fantasy/services/leagues';
+
+export function useMyLeagues(userId: string | undefined) {
+  return useQuery({
+    queryKey: ['fantasy-leagues', userId],
+    queryFn: () => getMyLeagues(userId!),
+    enabled: !!userId,
+    staleTime: 60_000,
+  });
+}
+
+export function useLeagueLeaderboard(leagueId: string | undefined) {
+  return useQuery({
+    queryKey: ['fantasy-league-leaderboard', leagueId],
+    queryFn: () => getLeagueLeaderboard(leagueId!),
+    enabled: !!leagueId,
+    staleTime: 60_000,
+  });
+}
