+++ src/features/fantasy/queries/keys.ts
@@ -1,0 +1,2 @@
+// Re-export from global factory for convenience within feature module
+export { qk } from '@/lib/queries/keys';
