+++ src/lib/queries/fantasyPicker.ts
@@ -1,26 +1,2 @@
-'use client';
-
-import { useQuery } from '@tanstack/react-query';
-import { qk } from './keys';
-import { getBatchFormScores } from '@/lib/services/scoring';
-import { getNextFixturesByClub } from '@/lib/services/fixtures';
-
-const FIVE_MIN = 5 * 60 * 1000;
-
-export function useBatchFormScores(playerIds: string[], enabled = true) {
-  return useQuery({
-    queryKey: [qk.scoring.batchForm, playerIds.length],
-    queryFn: () => getBatchFormScores(playerIds, 5),
-    enabled: enabled && playerIds.length > 0,
-    staleTime: FIVE_MIN,
-  });
-}
-
-export function useNextFixtures(enabled = true) {
-  return useQuery({
-    queryKey: qk.fixtures.next,
-    queryFn: getNextFixturesByClub,
-    enabled,
-    staleTime: FIVE_MIN,
-  });
-}
+// Re-export bridge — implementation moved to @/features/fantasy/queries/
+export * from '@/features/fantasy/queries/fantasyPicker';
